+++ src/beast/include/beast/core/detail/bind_handler.hpp
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2013-2016 Vinnie Falco (vinnie dot falco at gmail dot com)
+// Copyright (c) 2013-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
 //
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -42,7 +42,8 @@ public:
 
     template<class DeducedHandler>
     explicit
-    bound_handler(DeducedHandler&& handler, Args&&... args)
+    bound_handler(
+            DeducedHandler&& handler, Args&&... args)
         : h_(std::forward<DeducedHandler>(handler))
         , args_(std::forward<Args>(args)...)
     {
@@ -102,9 +103,11 @@ public:
 } // beast
 
 #include <functional>
+
 namespace std {
 template<class Handler, class... Args>
-void bind(beast::detail::bound_handler<
+void
+bind(beast::detail::bound_handler<
     Handler, Args...>, ...) = delete;
 } // std