+++ src/beast/include/beast/core/detail/is_call_possible.hpp
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2013-2016 Vinnie Falco (vinnie dot falco at gmail dot com)
+// Copyright (c) 2013-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
 //
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -26,8 +26,12 @@ std::false_type
 is_call_possible_test(C&& c, long, A&& ...a);
 
 /** Metafunction returns `true` if F callable as R(A...)
+
     Example:
+
+    @code
         is_call_possible<T, void(std::string)>
+    @endcode
 */
 /** @{ */
 template<class C, class F>
@@ -43,49 +47,6 @@ struct is_call_possible<C, R(A...)>
 {
 };
 /** @} */
-
-namespace test {
-
-struct is_call_possible_udt1
-{
-    void operator()(int) const;
-};
-
-struct is_call_possible_udt2
-{
-    int operator()(int) const;
-};
-
-struct is_call_possible_udt3
-{
-    int operator()(int);
-};
-
-#ifndef __INTELLISENSE__
-// VFALCO Fails to compile with Intellisense
-static_assert(is_call_possible<
-    is_call_possible_udt1, void(int)>::value, "");
-
-static_assert(! is_call_possible<
-    is_call_possible_udt1, void(void)>::value, "");
-
-static_assert(is_call_possible<
-    is_call_possible_udt2, int(int)>::value, "");
-
-static_assert(! is_call_possible<
-    is_call_possible_udt2, int(void)>::value, "");
-
-static_assert(! is_call_possible<
-    is_call_possible_udt2, void(void)>::value, "");
-
-static_assert(is_call_possible<
-    is_call_possible_udt3, int(int)>::value, "");
-
-static_assert(! is_call_possible<
-    is_call_possible_udt3 const, int(int)>::value, "");
-#endif
-
-} // test
 
 } // detail
 } // beast