+++ src/beast/include/beast/websocket/teardown.hpp
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2013-2016 Vinnie Falco (vinnie dot falco at gmail dot com)
+// Copyright (c) 2013-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
 //
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -16,7 +16,7 @@ namespace beast {
 namespace websocket {
 
 /** Tag type used to find @ref beast::websocket::teardown and @ref beast::websocket::async_teardown overloads
-    
+
     Overloads of @ref beast::websocket::teardown and
     @ref beast::websocket::async_teardown for user defined types
     must take a value of type @ref teardown_tag in the first