+++ src/beast/include/beast/zlib/detail/inflate_stream.hpp
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2013-2016 Vinnie Falco (vinnie dot falco at gmail dot com)
+// Copyright (c) 2013-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
 //
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -762,7 +762,7 @@ inflate_table(
     std::uint16_t offs[15+1];       // offsets in table for each length
 
     // Length codes 257..285 base
-    static std::uint16_t constexpr lbase[31] = { 
+    static std::uint16_t constexpr lbase[31] = {
         3, 4, 5, 6, 7, 8, 9, 10, 11, 13, 15, 17, 19, 23, 27, 31,
         35, 43, 51, 59, 67, 83, 99, 115, 131, 163, 195, 227, 258, 0, 0};