+++ src/beast/test/core/bind_handler.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2013-2016 Vinnie Falco (vinnie dot falco at gmail dot com)
+// Copyright (c) 2013-2017 Vinnie Falco (vinnie dot falco at gmail dot com)
 //
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -16,15 +16,17 @@ namespace beast {
 class bind_handler_test : public unit_test::suite
 {
 public:
-    static void foo (int)
+    void
+    callback(int v)
     {
+        BEAST_EXPECT(v == 42);
     }
 
     void run()
     {
-        auto f (bind_handler (
-            std::bind (&foo, std::placeholders::_1),
-            42));
+        auto f = bind_handler(std::bind(
+            &bind_handler_test::callback, this,
+                std::placeholders::_1), 42);
         f();
         pass();
     }